requirements.txt
ccxt>=4.4.0
pandas==2.2.2
ta==0.11.0
python-dotenv==1.0.1
aiosqlite==0.20.0
loguru==0.7.2
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
